requirements.txt
fastapi
uvicorn
openai
python-dotenv
streamlit
openai
python-dotenv
